pyproject.toml
[tool.poetry]
name = "prelovium"
version = "0.1.0"
description = "A web application for image processing and analysis"
authors = ["Valentin Buchner"]
readme = "README.md"
packages = [
    { include = "prelovium" }
]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.3.0"
numpy = "^1.26.4"
ipykernel = "^6.29.4"
opencv-python = "^4.9.0.80"
matplotlib = "^3.9.0"
transformers = "^4.41.0"
scikit-image = "^0.23.2"
torchvision = "^0.18.0"
ipywidgets = "^8.1.2"
google-cloud-aiplatform = "^1.51.0"
gradio = "^4.31.5"
flask = "^3.1.1"
gunicorn = "^21.2.0"
dotenv = "^0.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-flask = "^1.3.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.11.1"
black = "^23.7.0"
flake8 = "^6.0.0"
mypy = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
